Plugin handling requires setuptools at runtime.
This would hardly be an issue in any use case because setuptools
needs to be installed anyway for the build step. But without the
explicit dependency, the CI pipeline fails in the testing step:
https://dev.azure.com/conda-forge/feedstock-builds/_build/results?buildId=1050930&view=logs&j=7b6f2c87-f3a7-5133-8d84-7c03a75d9dfc&t=9eb77fd2-8ddd-5444-8fc0-71cb28dcb736&l=2623

In particular, plugins require checking entry points. That is done
by `pkg_resources` which is actually deprecated, so I put an
upper version boundary on the dependency.
Future versions of ODL should presumably migrate to `importlib` for
this functionality.